HACK: sm6150-common: Temporarily skip baseband check
With GLOBAL_THINLTO enabled, the updater binary fails to find
xiaomi.verify_baseband and we haven't figured a proper fix for that yet.

Since our device hasn't had an update for baseband in ages, we can
temporarily skip the check until we fix the real problem, but keep
global ThinLTO enabled for the performance gains.